+++ android/example/src/androidTest/java/com/example/ExampleInstrumentedTest.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication
+package com.example
 
 import androidx.test.ext.junit.runners.AndroidJUnit4
 import androidx.test.platform.app.InstrumentationRegistry
@@ -17,6 +17,6 @@ class ExampleInstrumentedTest {
     fun useAppContext() {
         // Context of the app under test.
         val appContext = InstrumentationRegistry.getInstrumentation().targetContext
-        assertEquals("com.example.myapplication", appContext.packageName)
+        assertEquals("com.example", appContext.packageName)
     }
 }
