+++ android/example/src/main/java/com/example/ToastPlugin.kt
@@ -1,4 +1,4 @@
-package com.example.myapplication
+package com.example
 
 import android.content.Context
 import android.widget.Toast
