+++ android/example/src/main/res/xml/file_paths.xml
@@ -1,4 +1,4 @@
 <?xml version="1.0" encoding="utf-8"?>
 <paths>
-    <external-path name="cache" path="Android/data/com.example.myapplication.debug/cache" />
+    <external-path name="cache" path="Android/data/com.example.debug/cache"/>
 </paths>
